requirements.txt
flask
yt-dlp
pyrogram
pytgcalls
tgcrypto
python-dotenv
